+++ javascript/src/ExpressionFactory.ts
@@ -1,6 +1,6 @@
-import CucumberExpression, { CucumberExpressionJson } from './CucumberExpression.js'
+import CucumberExpression from './CucumberExpression.js'
 import ParameterTypeRegistry from './ParameterTypeRegistry.js'
-import RegularExpression, { RegularExpressionJson } from './RegularExpression.js'
+import RegularExpression from './RegularExpression.js'
 import { Expression } from './types.js'
 
 export default class ExpressionFactory {
@@ -10,17 +10,5 @@ export default class ExpressionFactory {
     return typeof expression === 'string'
       ? new CucumberExpression(expression, this.parameterTypeRegistry)
       : new RegularExpression(expression, this.parameterTypeRegistry)
-  }
-
-  public createExpressionFromJson(json: CucumberExpressionJson | RegularExpressionJson) {
-    switch (json.type) {
-      case 'CucumberExpression':
-        return new CucumberExpression(json.expression, this.parameterTypeRegistry)
-      case 'RegularExpression':
-        return new RegularExpression(
-          new RegExp(json.expression, json.flags),
-          this.parameterTypeRegistry
-        )
-    }
   }
 }